% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{data.benchmark}
\alias{data.benchmark}
\title{MiRNA Sequencing Benchmark Data}
\format{
A data frame with 1033 rows and 54 columns. Here are the examples of the column and row naming rule:
\describe{
  \item{MXF2516_D13}{One sample belonging to MXF and library D with sample ID MXF2516 and library ID D13.}
  \item{hsa-let-7a-2*}{Gene ID.}
}
}
\usage{
data.benchmark
}
\description{
Myxofibrosarcoma (MXF) and pleomorphic malignant fibrous histiocytoma (PMFH) are the two most common and aggressive subtypes of genetically complex soft tissue sarcoma.
This dataset includes three libraries used for sequencing 54 individual tumor samples. Library preparation and read capture were each processed by a single experienced technician in one run.
}
\keyword{datasets}
