% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{data.group}
\alias{data.group}
\title{Sample Labels of the test and benchmark data}
\format{
A set of labels for the test and benchmark samples, either "MXF" (myxofibrosarcoma) or "PMFH" (pleomorphic malignant fibrous histiocytoma).
}
\usage{
data.group
}
\description{
Sample Labels of the test and benchmark data
}
\keyword{datasets}
