% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{data.miR.info}
\alias{data.miR.info}
\title{MiRNA Information}
\format{
Information of the microRNAs in test and benchmark dataset, including their names, exact sequence, length and GC content.
}
\usage{
data.miR.info
}
\description{
MiRNA Information
}
\keyword{datasets}
