% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFunc.R
\name{DE.statistics}
\alias{DE.statistics}
\title{Statistics for DEA Results Based Golden Standards}
\usage{
DE.statistics(markers, id.list, truth, selected.marker = NULL)
}
\arguments{
\item{markers}{Vector of all markers considered in the analysis.}

\item{id.list}{Vector of markers that are differentially expressed in a
data set --- typically in the data set \code{\link{data.test}}.
Differential expression can be computed via \code{\link{DE.edgeR}}
or \code{\link{DE.voom}}.}

\item{truth}{Vector of genes that are (assumedly) truly differential
expressed, e.g. basen on a gold standard --- typically the data set
\code{\link{data.benchmark}}.}

\item{selected.marker}{optional Vector of a subset of \code{markers}.
If given, the analysis will be limited to the given subset.
Leave \code{NULL} if all markers are considered for the analysis.}
}
\value{
A list of:
\describe{
  \item{TPR}{True positive Rate}
  \item{FPR}{False Positive Rate}
  \item{FDR}{False Discovery Rate}
  \item{FNR}{False Negative Rate}
}
}
\description{
Computes the agreement of differential expression (DE) wihtin one data set
with an assumed true DE status usually given based on a gold standard.
The following statistics are computed based of the differentially expressed
markers \code{id.list} compared to the list truly differentially expressed
markers \code{truth}:
\itemize{
  \item False Discovery Rate (FDR)
  \item False Negative Rate (FNR)
  \item True Positive Rate (TPR)
  \item False Positive Rate (FPR)
}
}
\details{
Both, \code{id.list} and \code{truth} are a subset of markers from
\code{markers}.
}
\examples{
DE.bench <- DE.voom(data.benchmark, data.group)
DE.test <- DE.voom(data.test, data.group)
stats <- DE.statistics(rownames(data.benchmark), DE.test$id.list, DE.bench$id.list)
print(stats)
}
