% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFunc.R
\name{DE.voom}
\alias{DE.voom}
\title{Differential Expression Analysis Using Voom-limma Pipeline}
\usage{
DE.voom(RC, groups, Pval = 0.01, normalized = TRUE, adjust = NULL)
}
\arguments{
\item{RC}{Data in the format of a data frame or matrix, with columns for samples and rows for genes.}

\item{groups}{Vector of characters indicating the group label for each sample.}

\item{Pval}{Cut-off point for p-values to identify differentially expressed genes.}

\item{normalized}{Logical, whether the data is provided as normalized counts.
If set to FALSE, adjustment factors must be provided using the \code{adjust}
parameter.}

\item{adjust}{Adjusting factors for normalizing the count data. Must be
provided if the data is not normalized beforehand as indicated by the
\code{normalized} parameter.}
}
\value{
list, containing \code{id.list} (names of DE genes), \code{p.val}, and \code{log2.FC}.
}
\description{
Perform DEA using the voom-limma pipeline on a normalized dataset.
The normalized data can be provided as normalized counts or by adjusting
factor for the original count data.
}
\examples{
voom.benchmark <- DE.voom(data.benchmark, data.group)
str(voom.benchmark)
}
\references{
\href{https://ucdavis-bioinformatics-training.github.io/2018-June-RNA-Seq-Workshop/thursday/DE.html}{Differential Expression with Limma-Voom}
}
