% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.dendrogram}
\alias{fig.dendrogram}
\title{Dendrogram for clustering p-values}
\usage{
fig.dendrogram(
  MethodsCompare,
  DE.method = "DE.voom",
  QN_filter = FALSE,
  Pval = 0.01,
  MethodNew,
  pvalues,
  Methods_visual = c(MethodsCompare, MethodNew),
  test = data.test,
  group = data.group
)
}
\arguments{
\item{MethodsCompare}{the vector of methods that researchers would like to compare with, and selected from  \code{norm.none}, \code{norm.TMM}, \code{norm.TC}, \code{norm.UQ}, \code{norm.med}, \code{norm.DESeq}, \code{norm.PoissonSeq}, \code{norm.QN}, \code{norm.SVA}, and \code{norm.RUV}.}

\item{DE.method}{the DEA method selected from \code{DE.voom} and \code{DE.edgeR}, and default to be \code{DE.voom}.}

\item{QN_filter}{whether the filtering is performed if \code{norm.QN} is used.}

\item{MethodNew}{the name of the new method which will be presented in the legend.}

\item{pvalues}{a vector of p values computed from DE analysis using the new method normalized test data.}

\item{test}{test data with the same format of data.test, which are used for test p-values, and is set to use data.test by default}

\item{group}{vector of characters indicating the group for each sample (only 2 groups allowed).}
}
\value{
figure of dendrogram
}
\description{
Function for clustering normalization methods based on the p-values pattern calculated from the same dataset.
}
\examples{
t <-  runif(1033)
names(t) <-  rownames(data.test)
fig.dendrogram(MethodsCompare = c("norm.none", "norm.TMM", "norm.SVA", "norm.TC", "norm.RUVr"), MethodNew = "Example", pvalues = t)
}
