% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.RLE}
\alias{fig.RLE}
\title{Relative Log Expression Plot}
\usage{
fig.RLE(data, groups, title)
}
\arguments{
\item{data}{Raw or normalized count data}

\item{groups}{vector of characters indicating the group for each sample
(only 2 groups are allowed).}

\item{title}{optional Figure title}
}
\value{
ggplot boxplot for relative log expression
}
\description{
Function for generating relative log expression plots based on raw or
normalized count data as the input.
}
\examples{
fig.RLE(data.test, data.group, title="test data (no normalization)")
}
