% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.venn}
\alias{fig.venn}
\title{Venn diagram for p-values}
\usage{
fig.venn(benchmark.pval, test.pval, Pvalue)
}
\arguments{
\item{benchmark.pval}{p-values obtained from benchmark data}

\item{test.pval}{p-values obtained from normalized test data}

\item{Pvalue}{Cut-off point for p-values for identifying significance}
}
\value{
A Venn diagram
}
\description{
Venn diagram is used to identify the performance of different normalization methods based on intersection of differential expressed genes.
}
\examples{
benchmark.voom <- DE.voom(RC = data.benchmark, groups = data.group, P = 0.01)
test.voom <- DE.voom(RC = data.test, groups = data.group, P = 0.01)
fig.venn(benchmark.voom$p.val, test.voom$p.val, 0.01)
}
