% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.volcano}
\alias{fig.volcano}
\title{Volcano Figure}
\usage{
fig.volcano(normtest, groups, adjust = NULL)
}
\arguments{
\item{normtest}{normalized data using the method which the researchers desire to compare with the other methods, or it could be the raw data if the adjusting parameters available.}

\item{groups}{vector of characters indicating the group for each sample.}

\item{adjust}{adjusting parameters for each sample.}
}
\value{

}
\description{
Function for generating volcano figures with log fold change as x-axis and -log10(p-value) as y-axis.
}
\examples{
fig.volcano(benchmark, group)
}
