% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombAnalysis.R
\name{full.pipeline}
\alias{full.pipeline}
\title{Full pipeline function}
\usage{
full.pipeline(
  bench.data,
  test.data,
  groups.data,
  norm.counts,
  adjust.factors,
  method.name,
  DE.method,
  Pval,
  generate.plots
)
}
\description{
This internal function performs the full precision.seq pipeline.
It is used by the functions \code{\link{precision.seq}} and
\code{\link{pip.simulated.data}}.
}
\keyword{internal}
