% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{miR.info}
\alias{miR.info}
\title{MiRNA Information}
\format{Information of the microRNAs in test and benchmark dataset, including their names, exact sequence, length and GC content.}
\usage{
miR.info
}
\description{
MiRNA Information
}
\keyword{datasets}
