% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFunc.R
\name{norm.DESeq}
\alias{norm.DESeq}
\title{Normalization By DESeq (DESeq)}
\usage{
norm.DESeq(raw, groups)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and rows for genes.}

\item{groups}{vector of characters indicating the group for each sample.}
}
\value{
list, containing \code{dat.normed} (normalized dataset) and \code{scaling.factor} (scaling factor) for each sample.
}
\description{
Normalize the dataset using DESeq, and return the normalized dataset with scaling factor.
}
\examples{
test.DESeq <- norm.DESeq(data.test, data.group)
}
\references{
\href{https://bioconductor.org/packages/release/bioc/vignettes/DESeq/inst/doc/DESeq.pdf}{DESeq Package}
}
