% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFunc.R
\name{norm.PoissonSeq}
\alias{norm.PoissonSeq}
\title{Normalization By PoissonSeq (PoissonSeq)}
\usage{
norm.PoissonSeq(raw)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and raws for genes.}
}
\value{
list, containing \code{dat.normed} (normalized dataset) and \code{scaling.factor} (scaling factor) for each sample.
}
\description{
Normalize the dataset using PoissonSeq, and return the normalized dataset with scaling factor.
}
\examples{
test.PoissonSeq <- norm.PoissonSeq(data.test)
}
\references{
\href{https://cran.r-project.org/web/packages/PoissonSeq/PoissonSeq.pdf}{PossionSeq Package}
}
