% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFunc.R
\name{norm.QN}
\alias{norm.QN}
\title{Normalization By Quantile Normalization (QN)}
\usage{
norm.QN(raw, filter = FALSE)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{filter}{whether filter before normalization}
}
\value{
list, containing \code{dat.normed} (normalized dataset).
}
\description{
Normalize the dataset using QN, and return the normalized dataset with scaling factor.
}
\examples{
test.QN <- norm.QN(data.test)
}
\references{
\href{http://jtleek.com/genstats/inst/doc/02_05_normalization.html}{Quantile Normalization Tutorial}
}
