% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFunc.R
\name{norm.RUVg}
\alias{norm.RUVg}
\title{Normalization By Remove Unwanted Variation Using Control Genes (RUVg)}
\usage{
norm.RUVg(raw, groups)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample.}
}
\value{
list, containing \code{dat.normed} (normalized dataset), and the \code{adjust.factor} (adjusting factors) for the design matrix. The normalized dataset could only used for exploration, and adjusting factors are recommended as a covariate in the downstream analysis.
}
\description{
Normalize the dataset using RUV Using Control Genes, and return the normalized dataset with adjusting factor.
}
\examples{
test.RUVg <- norm.RUV(data.test, data.group)
}
\references{
\href{http://www.bioconductor.org/packages/devel/bioc/vignettes/RUVSeq/inst/doc/RUVSeq.pdf}{RUVSeq Tutorial}
}
