% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFunc.R
\name{norm.SVA}
\alias{norm.SVA}
\title{Normalization By Surrogate Variable Analysis for Sequencing Data (SVA)}
\usage{
norm.SVA(raw, groups)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample.}

\item{calibrator}{whether the raw dataset includes the known negative control genes which should contain "cali" in the gene names.}
}
\value{
list, containing \code{dat.normed} (normalized dataset) without the calibrators, and the \code{adjust.factor} (adjusting factors) for the design matrix. The normalized dataset could only used for exploration.
}
\description{
Normalize the dataset using SVA, and return the normalized dataset with adjusting factor.
}
\examples{
group = c(rep(c(rep('MXF',9),rep('PMFH',9)),3))
test.SVA = norm.SVA(test, group)
}
\references{
\href{https://bioconductor.org/packages/release/bioc/vignettes/sva/inst/doc/sva.pdf}{SVA Package}
}
