% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFunc.R
\name{norm.TC}
\alias{norm.TC}
\title{Normalization By Total Count (TC)}
\usage{
norm.TC(raw, groups)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample.}
}
\value{
list, containing \code{dat.normed} (normalized dataset) and \code{scaling.factor} (scaling factor) scaled by 1e6 for each sample.
}
\description{
Normalize the dataset using TC, and return the normalized dataset with scaling factor.
}
\examples{
test.TC <-  norm.TC(data.test, data.group)
}
\references{
Dillies, Marie-Agnès, et al. "A comprehensive evaluation of normalization methods for Illumina high-throughput RNA sequencing data analysis." \emph{Briefings in bioinformatics} 14.6 (2013): 671-683.
}
