% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombAnalysis.R
\name{pip.fig}
\alias{pip.fig}
\title{Pipeline for visualization Using}
\usage{
pip.fig(
  norm.method,
  QN_filter = FALSE,
  new.method.name,
  normalized.test = NULL,
  adjust.factor = NULL,
  DE.method = "DE.voom",
  Pval_cutoff = 0.01
)
}
\arguments{
\item{norm.method}{the normalization method selected for comparsion from \code{norm.none}, \code{norm.TMM}, \code{norm.TC}, \code{norm.UQ}, \code{norm.med}, \code{norm.DESeq}, \code{norm.PoissonSeq}, \code{norm.QN}, \code{norm.SVA}, and \code{norm.RUV}.}

\item{QN_filter}{whether the filtering is performed if \code{norm.QN} selected}

\item{new.method.name}{the name of new normalization method}

\item{normalized.test}{the normalized test data by the new normalization method}

\item{adjust.factor}{the adjusting factors for the test data by the new normalization method}

\item{DE.method}{the method for differential expression analysis from \code{DE.voom} and \code{DE.edgeR}, default to be \code{DE.voom}}

\item{Pval_cutoff}{p-value for identifying DE genes, default to be 0.01}
}
\value{

}
\description{
A pipeline for visualization to compare the normalization method with selected existing methods
}
\examples{
vsn.norm <- vsn::justvsn(data.test)
vsn.norm <- ifelse(vsn.norm<0, 1, vsn.norm)
pip.fig(c("norm.none", "norm.TMM", "norm.SVA", "norm.TC", "norm.RUVr"),
 new.method.name = "vsn", normalized.test = vsn.norm)
}
