% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombAnalysis.R
\name{pip.simulated.data}
\alias{pip.simulated.data}
\title{Full normalization assessment for simulated data}
\usage{
pip.simulated.data(
  data,
  groups,
  norm.counts,
  adjust.factors,
  method.name,
  DE.method = "DE.voom",
  Pval = 0.01
)
}
\arguments{
\item{data}{list of paired data sets. Each pair must consist of
a benchmark data set "simulated_benchmark" and a test data set
"simulated_test".}

\item{groups}{Sample groups in the simulated test and benchmark data sets.
Must be the same across all pairs.}

\item{norm.counts}{Normalized counts of each simulated test data set for a
user-provided normalization method.}

\item{adjust.factors}{optional Adjustment factors for the normalization of the
test data for a user-provided normalization method.}

\item{method.name}{optional Name of the normalization method. Used for
naming of data frames and plot legends.}

\item{DE.method}{Method for computing differential expression statuses.
Available: "DE.voom" and "DE.edgeR".}

\item{Pval}{P-value cutoff for differential expression.}
}
\value{
List of assessment statistics and plots (ggplot objects):
\describe{
  \item{DE.stats}{List of statistics (true positive rate, false positive
  rate, false discovery rate, and false negative rate) based on the
  comparison of differential expression statuses between the normalized test
  data and the unnormalized gold standard (the benchmark data).
  Contains results for each pair of simulated data sets.}
  \item{p.boxplot}{Boxplot for False Negative Rate (FNR) and False
  Discovery Rate (FDR) of the agreement of DE statuses between the simulated
  test and benchmark data sets.}
}
}
\description{
Simulated data can be sampled using the function \code{\link{simulated.data}}.
}
