% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{plot.CAT}
\alias{plot.CAT}
\title{Concordance At The Top Plot}
\usage{
\method{plot}{CAT}(MethodType, MethodName, pvalues)
}
\arguments{
\item{MethodType}{the type of method that researchers would like to compare with, either "Scale" or "Regression"}

\item{MethodName}{the name of the new method which will be presented in the legend}

\item{pvalues}{a vector of p values computed from DE analysis using the normalized data}
}
\value{

}
\description{
Function for generating concordance at the top plot, which compares concordance of the p-values obtained
from benchmark data without normalization and normalized test data.
}
\examples{
t = runif(1033)
names(t) = rownames(test)
plot.CAT("Scale", "Example", t)

}
