% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulation.algorithm}
\alias{simulation.algorithm}
\title{The Algorithm for Obtaining the Simulation Plan
This algorithm is tailored for the data.benchmark included in this package.}
\usage{
simulation.algorithm(benchmark)
}
\arguments{
\item{benchmark}{the data.benchmark used for the simulation}
}
\value{

}
\description{
The Algorithm for Obtaining the Simulation Plan
This algorithm is tailored for the data.benchmark included in this package.
}
\examples{
simulation.plan <- simulation.algorithm(data.benchmark)
}
