% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulation Plan}
\format{A dataframe including information of allocation about 20,000 simulation datasets, which contains the proportion of differential expression and median of mean difference for each dataset.}
\usage{
simulation
}
\description{
Simulation Plan
}
\keyword{datasets}
