% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFunc.R
\name{DE.edgeR}
\alias{DE.edgeR}
\title{Differential Expression Analysis Using EdgeR}
\usage{
DE.edgeR(RC, groups, Pval = 0.01, normalized = TRUE, adjust = NULL)
}
\arguments{
\item{RC}{data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample.}

\item{Pval}{cut-off point for p-values to identify DE genes.}

\item{normalized}{Logical, whether the data is provided as normalized counts.
If set to FALSE, adjustment factors must be provided using the \code{adjust}
parameter.}

\item{adjust}{Adjusting factors for normalizing the count data. Must be
provided if the data is not normalized beforehand as indicated by the
\code{normalized} parameter.}
}
\value{
list, containing \code{id.list} (names of DE genes), \code{p.val}, and \code{log2.FC}.
}
\description{
Perform DEA by EdgeR on normalized dataset or dataset with adjusting factor.
}
\examples{
edgeR.benchmark <- DE.edgeR(data.benchmark, data.group)
}
\references{
\url{https://ucdavis-bioinformatics-training.github.io/2018-June-RNA-Seq-Workshop/thursday/DE.html}
}
