% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFunc.R
\name{DE.edgeR}
\alias{DE.edgeR}
\title{Differential Expression Analysis Using EdgeR}
\usage{
DE.edgeR(RC, groups, P = 0.01, normalized = TRUE, adjust = NULL)
}
\arguments{
\item{RC}{data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample.}

\item{P}{cut-off point for p-values to identify DE genes.}

\item{normalized}{whether the dataset is normalized (by scaling method)}

\item{adjust}{the adjusting factor if the dataset itself is not normalized}
}
\value{
list, containing \code{id.list} (names of DE genes) and \code{p.val} (corresponding p-values and log2-fold-change for all genes).
}
\description{
Perform DEA by EdgeR on normalized dataset or dataset with adjusting factor.
}
\examples{
group = c(rep(c(rep('MXF',9),rep('PMFH',9)),3))
edgeR.benchmark = DE.edgeR(benchmark, group)
}
\references{
\url{https://ucdavis-bioinformatics-training.github.io/2018-June-RNA-Seq-Workshop/thursday/DE.html}
}
