% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFunc.R
\name{DE.statistics}
\alias{DE.statistics}
\title{Statistics for DEA Results Based Golden Standards}
\usage{
DE.statistics(markers, id.list, truth, selected.marker = NULL)
}
\arguments{
\item{id.list}{vector of genes that are differentially expressed given a DEA
using voom or edgeR.}

\item{truth}{vector of genes that are truly differential expressed}

\item{selected.marker}{if given, vector of a subset of genes/markers for
this analysis. Leave \code{NULL} if all markers are considered for the analysis.}

\item{merkers}{vector of all markers considered in the analysis.}
}
\value{
a list of values about TPR, FPR, FDR, FNR
}
\description{
Computing the true positive rate, false positive rate, false discovery rate, and false negative rate based on given golden standards.
}
