% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFunc.R
\name{DE.voom}
\alias{DE.voom}
\title{Differential Expression Analysis Using Voom-limma Pipeline}
\usage{
DE.voom(RC, groups, Pval = 0.01, normalized = TRUE, adjust = NULL)
}
\arguments{
\item{RC}{data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample.}

\item{Pval}{cut-off point for p-values to identify DE genes.}

\item{normalized}{whether the dataset is normalized (by scaling method)}

\item{adjust}{the adjusting factor if the dataset itself is not normalized}
}
\value{
list, containing \code{id.list} (names of DE genes), \code{p.val}, and \code{log2.FC}.
}
\description{
Perform DEA by voom-limma pipeline on normalized dataset or dataset with adjusting factor.
}
\examples{
voom.benchmark <- DE.voom(data.benchmark, data.group)
}
\references{
\href{https://ucdavis-bioinformatics-training.github.io/2018-June-RNA-Seq-Workshop/thursday/DE.html}{Differential Expression with Limma-Voom}
}
