% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.CAT}
\alias{fig.CAT}
\title{Concordance At The Top Plot}
\usage{
fig.CAT(DEA, truth.DEA, title, maxrank = 100, subset = NULL)
}
\arguments{
\item{DEA}{Results of a differential expression analysis using the
included methods \code{\link{DE.voom}} or \code{\link{DE.edgeR}}.
Must be a list including
\describe{
  \item{id.list}{List of differentially expressed markers}
  \item{p.val}{p-values for differential expression}
}}

\item{truth.DEA}{Gold standard (assumed truth) for differential expression.
Must be in the same format as \code{DEA}.}

\item{title}{optional Figure title}

\item{maxrank}{optional specify the maximum size of top-ranked items that you want to plot.}

\item{subset}{optional vector of a subset of genes/markers for this analysis}
}
\value{
figure of concordance for comparison
}
\description{
Function for generating a concordance at the top plot, which compares
concordance of the p-values of differential expression --- typically from the
normalized or raw data set \code{\link{data.test}} --- to the p-values of
differential expression of a gold standard --- typically
the data set \code{\link{data.benchmark}}.
}
