% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.CAT}
\alias{fig.CAT}
\title{Concordance At The Top Plot}
\usage{
fig.CAT(DEA, truth.DEA, title, maxrank = 100, subset = NULL)
}
\arguments{
\item{DEA}{a list of differential expression analysis results with the element names to be the normalization methods}

\item{truth.DEA}{differential expression analysis results from the benchmark (gold standard) obtained from DE.voom, DE.edge, or any method from the users by storing the results as same as DE methods in the package (including DE genes, p-values and log2 fold changes)}

\item{title}{figure title}

\item{maxrank}{optionally specify the maximum size of top-ranked items that you want to plot.}

\item{subset}{vector of a subset of genes/markers for this analysis}
}
\value{
figure of concordance for comparison
}
\description{
Function for generating concordance at the top plot, which compares concordance of the p-values obtained
from benchmark data without normalization and normalized test data.
}
