% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.CAT}
\alias{fig.CAT}
\title{Concordance At The Top Plot}
\usage{
fig.CAT(DEA, truth.DEA, title, maxrank = 100, subset = NULL)
}
\arguments{
\item{DEA}{Results of a differential expression analysis using the
included methods \code{\link{DE.voom}} or \code{\link{DE.edgeR}}.
Must be a list including
\describe{
  \item{id.list}{List of differentially expressed markers}
  \item{p.val}{p-values for differential expression}
}}

\item{truth.DEA}{Gold standard (assumed truth) for differential expression.
Must be in the same format as \code{DEA}.}

\item{title}{optional Figure title}

\item{maxrank}{optional specify the maximum size of top-ranked items that you want to plot.}

\item{subset}{optional vector of a subset of genes/markers for this analysis}
}
\value{
figure of concordance for comparison
}
\description{
Function for generating a concordance at the top plot, which compares
concordance of the p-values of differential expression --- typically from the
normalized or raw data set \code{\link{data.test}} --- to the p-values of
differential expression of a gold standard --- typically
the data set \code{\link{data.benchmark}}.
}
\examples{
voom.benchmark <- DE.voom(data.benchmark, data.group)
test.norm <- pip.norm(raw=data.test, groups=data.group, norm.method = "all")
test.DE <- list(
TMM = DE.voom(RC=test.norm$TMM$dat.normed, groups = data.group),
TC = DE.voom(RC=test.norm$TC$dat.normed, groups = data.group),
UQ = DE.voom(RC=test.norm$UQ$dat.normed, groups = data.group),
med = DE.voom(RC=test.norm$med$dat.normed, groups = data.group),
DESeq = DE.voom(RC=test.norm$DESeq$dat.normed, groups = data.group),
PoissonSeq = DE.voom(RC=test.norm$PoissonSeq$dat.normed, groups = data.group),
QN = DE.voom(RC=test.norm$QN$dat.normed, groups = data.group),
RUVg = DE.voom(RC=data.test, groups = data.group, normalized=FALSE, adjust=test.norm$RUVg$adjust.factor),
RUVs = DE.voom(RC=data.test, groups = data.group, normalized=FALSE, adjust=test.norm$RUVs$adjust.factor),
RUVr = DE.voom(RC=data.test, groups = data.group, normalized=FALSE, adjust=test.norm$RUVr$adjust.factor),
SVA = DE.voom(RC=data.test, groups = data.group, normalized=FALSE, adjust=test.norm$SVA$adjust.factor),
noNorm = DE.voom(RC=data.test, groups = data.group))

fig.CAT(DEA = test.DE, truth.DEA = voom.benchmark, title = "Example of CAT plot")
}
