% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.dendrogram}
\alias{fig.dendrogram}
\title{Dendrogram for clustering p-values}
\usage{
fig.dendrogram(DEA.pval.list, title, subset = NULL)
}
\arguments{
\item{DEA.pval.list}{A list of p-values from differential expression analysis results with the element names to be the normalization methods}

\item{title}{optional Figure title}

\item{subset}{optional Vector of a subset of markers.
If given, the dendrogram analysis will be limited to the given subset
of markers. Leave \code{NULL} if all markers should be considered.}
}
\value{
Figure of dendrogram
}
\description{
Function for clustering normalization methods based on the p-values pattern calculated from the same dataset.
}
\examples{
t <-  runif(1033)
names(t) <-  rownames(data.test)
fig.dendrogram(MethodsCompare = c("norm.none", "norm.TMM", "norm.SVA", "norm.TC", "norm.RUVr"), MethodNew = "Example", pvalues = t)
}
