% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.dendrogram}
\alias{fig.dendrogram}
\title{Dendrogram for clustering p-values}
\usage{
fig.dendrogram(DEA, title, subset = NULL)
}
\arguments{
\item{DEA}{a list of differential expression analysis results with the element names to be the normalization methods}

\item{title}{figure title}

\item{subset}{vector of a subset of genes/markers for this analysis}
}
\value{
figure of dendrogram
}
\description{
Function for clustering normalization methods based on the p-values pattern calculated from the same dataset.
}
\examples{
t <-  runif(1033)
names(t) <-  rownames(data.test)
fig.dendrogram(MethodsCompare = c("norm.none", "norm.TMM", "norm.SVA", "norm.TC", "norm.RUVr"), MethodNew = "Example", pvalues = t)
}
