% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.FDR_FNR.boxplot}
\alias{fig.FDR_FNR.boxplot}
\title{Boxplot of FDR and FNR for Simulated data}
\usage{
fig.FDR_FNR.boxplot(DEA.list, truth.DEA.list, title, subset = NULL)
}
\arguments{
\item{DEA.list}{A list with each element as a normalization method including the differential expression analysis (DEA) results from multiple simulated data}

\item{truth.DEA.list}{A list of DEA results of benchmark data for multiple simulated dataset}

\item{title}{The title of the figure}

\item{subset}{optional A vector of a subset of genes/markers for this analysis}
}
\value{

}
\description{
Function for the boxplots of FDR and FNR based on golden truth (typically the simulated benchmark data set \code{\link{simulated.data()}})
and the differential expression analysis (typically the results from the simulated normalized test data set \code{\link{simulated.data()}}).
}
