% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.FDR_FNR}
\alias{fig.FDR_FNR}
\title{Selection of normalization methods based on golden standards}
\usage{
fig.FDR_FNR(DEA, truth.DEA, title, subset = NULL)
}
\arguments{
\item{DEA}{a list of differential expression analysis results with the element names to be the normalization methods}

\item{truth.DEA}{differential expression analysis results from the benchmark (gold standard) obtained from DE.voom, DE.edge, or any method from the users by storing the results as same as DE methods in the package (including DE genes, p-values and log2 fold changes)}

\item{title}{figure title}

\item{subset}{vector of a subset of genes/markers for this analysis}
}
\value{
figure for selection of normalization methods
}
\description{
Selection of normalization methods based on golden standards
}
