% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.RLE}
\alias{fig.RLE}
\title{Relative Log Expression Plot}
\usage{
fig.RLE(data, groups, title)
}
\arguments{
\item{data}{normalized testing using the method which the researchers desire to compare with the other methods}

\item{groups}{vector of characters indicating the group for each sample (only 2 groups are allowed).}

\item{title}{the title of the figure}
}
\value{
boxplot for relative log expression
}
\description{
Function for generating relative log expression plot using the read count data as the input.
}
\examples{
fig.RLE(data.test, data.group, "test without normalization")
}
