% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.RLE}
\alias{fig.RLE}
\title{Relative Log Expression Plot}
\usage{
fig.RLE(normtest, main)
}
\arguments{
\item{normtest}{normalized testing using the method which the researchers desire to compare with the other methods}

\item{main}{the title of the figure}
}
\value{
boxplot for relative log expression
}
\description{
Function for generating relative log expression plot using the normalized test data as the input.
The order of the columns in the normalized test data is transformed to be as same as the sorted benchmark data, which makes the figure comparable to the other methods.
}
\examples{
fig.RLE(data.test, "test without normalization")
}
