% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunc.R
\name{fig.volcano}
\alias{fig.volcano}
\title{Volcano Figure}
\usage{
fig.volcano(DEA.res, title)
}
\arguments{
\item{DEA.res}{the list of differential expression analysis results obtained from DE.voom, DE.edge, or any method from the users by storing the results as same as DE methods in the package (including DE genes, p-values and log2 fold changes).}

\item{title}{the title of the figure}
}
\value{
volcano plot
}
\description{
Function for generating volcano figures with log fold change as x-axis and -log10(p-value) as y-axis.
}
\examples{
voom.benchmark <- DE.voom(data.benchmark, data.group)
fig.volcano(voom.benchmark, title = "Volcano Plot")
}
