% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFunc.R
\name{norm.TMM}
\alias{norm.TMM}
\title{Normalization By Trimmed Mean of M-values (TMM)}
\usage{
norm.TMM(raw, groups)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample.}
}
\value{
list, containing \code{dat.normed} (normalized dataset) and \code{scaling.factor} (scaling factor) for each sample.
}
\description{
Normalize the dataset using TMM, and return the normalized dataset with scaling factor.
}
\examples{
test.TMM <-  norm.TMM(data.test, data.group)
}
\references{
\href{https://www.bioconductor.org/packages/devel/bioc/vignettes/edgeR/inst/doc/edgeRUsersGuide.pdf}{edgeR User Guide}
}
