% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombAnalysis.R
\name{pip.norm}
\alias{pip.norm}
\title{Normalization for RNASeq Data}
\usage{
pip.norm(raw, groups, norm.method, QN_filter = FALSE)
}
\arguments{
\item{raw}{raw count data in the format of data frame or matrix, with columns for samples and raws for genes.}

\item{groups}{vector of characters indicating the group for each sample (only 2 groups allowed).}

\item{norm.method}{the method for normalization selected from \code{norm.none}, \code{norm.TMM}, \code{norm.TC}, \code{norm.UQ}, \code{norm.med}, \code{norm.DESeq}, \code{norm.PoissonSeq}, \code{norm.QN}, \code{norm.SVA}, \code{norm.RUVr}, \code{norm.RUVg}, and \code{norm.RUVs}.}

\item{QN_filter}{whether the filtering is performed if \code{method = norm.QN}.}
}
\value{
list, containing normalized dataset, and scaling factors or adjusting factors.
}
\description{
Normalization for RNASeq Data
}
\examples{
res <- pip.norm(data.test, data.group, "norm.TMM")
}
