% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombAnalysis.R
\name{precision.seq}
\alias{precision.seq}
\title{Full normalization assessment for given normalized test data}
\usage{
precision.seq(
  norm.counts,
  adjust.factors,
  method.name,
  DE.method = "DE.voom",
  Pval = 0.01
)
}
\arguments{
\item{normalized.test}{}
}
\value{

}
\description{
Full normalization assessment for given normalized test data
}
