% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombAnalysis.R
\name{precision.seq}
\alias{precision.seq}
\title{Full normalization assessment for given normalized test data}
\usage{
precision.seq(normalized.test)
}
\arguments{
\item{normalized.test}{}
}
\value{

}
\description{
Full normalization assessment for given normalized test data
}
